coding-challenges/cracking-the-coding-interview/problems/07_rotateMatrix.ts
// 7. *Rotate Matrix*:

// Given an image represented by an NxN matrix, where each pixel in the image is 4
// bytes, write a method to rotate the image by 90 degrees. Can you do this in place?

type Matrix = number[][]

export default function rotateMatrix (matrix: Matrix) {
  let left = 0;
  let right = matrix.length - 1;

  while (left < right) {
    for (let i = 0; i < right - left; i++) {
      const top = left
      const bottom = right
      const endAtTopLeft = matrix[bottom - i][left]

      matrix[bottom - i][left] = matrix[bottom][right - i]
      matrix[bottom][right - i] = matrix[top + i][right]
      matrix[top + i][right] = matrix[top][left + i]
      matrix[top][left + i] = endAtTopLeft
    }

    left += 1
    right -= 1
  }
}
